improved driving mode detection

+++ Inc/util.h
@@ -110,8 +110,8 @@ void sideboardSensors(uint8_t sensors);
 
 #ifdef VARIANT_BBCAR
   void adc_error_melody();
-  int32_t isAroundMin(int16_t value, int16_t minn, int16_t maxx);
-  int32_t isAroundMax(int16_t value, int16_t minn, int16_t maxx);
+  int32_t isAroundMin(int32_t value, int32_t minn, int32_t maxx);
+  int32_t isAroundMax(int32_t value, int32_t minn, int32_t maxx);
   void bbcarDetectDrivingMode();
   int16_t bbcarLoop();
 #endif
